Body segments where not spawning, it was a problem where the food dissapeared before being compared in the body script, changed functions from FixedUpdate to Update
