+++ Assets/Scenes/Snake.unity
@@ -180,11 +180,6 @@ BoxCollider2D:
   serializedVersion: 2
   m_Size: {x: 1, y: 1}
   m_EdgeRadius: 0
---- !u!1 &406708139 stripped
-GameObject:
-  m_CorrespondingSourceObject: {fileID: 4780089334309980615, guid: 160d10cd78b53cb4081a88afe75f72f9, type: 3}
-  m_PrefabInstance: {fileID: 4780089334442481260}
-  m_PrefabAsset: {fileID: 0}
 --- !u!4 &406708141 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 4780089334309980609, guid: 160d10cd78b53cb4081a88afe75f72f9, type: 3}
@@ -448,7 +443,7 @@ PrefabInstance:
   m_ObjectHideFlags: 0
   serializedVersion: 2
   m_Modification:
-    m_TransformParent: {fileID: 1133086475}
+    m_TransformParent: {fileID: 0}
     m_Modifications:
     - target: {fileID: 4056836922415472164, guid: 5683d367fe60ae24586720bcd5a4aed1, type: 3}
       propertyPath: m_Name
@@ -456,7 +451,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 4056836922415472167, guid: 5683d367fe60ae24586720bcd5a4aed1, type: 3}
       propertyPath: m_RootOrder
-      value: 0
+      value: 7
       objectReference: {fileID: 0}
     - target: {fileID: 4056836922415472167, guid: 5683d367fe60ae24586720bcd5a4aed1, type: 3}
       propertyPath: m_LocalPosition.x
@@ -468,7 +463,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 4056836922415472167, guid: 5683d367fe60ae24586720bcd5a4aed1, type: 3}
       propertyPath: m_LocalPosition.z
-      value: 0
+      value: -0.1772651
       objectReference: {fileID: 0}
     - target: {fileID: 4056836922415472167, guid: 5683d367fe60ae24586720bcd5a4aed1, type: 3}
       propertyPath: m_LocalRotation.w
@@ -476,15 +471,15 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 4056836922415472167, guid: 5683d367fe60ae24586720bcd5a4aed1, type: 3}
       propertyPath: m_LocalRotation.x
-      value: 0
+      value: -0
       objectReference: {fileID: 0}
     - target: {fileID: 4056836922415472167, guid: 5683d367fe60ae24586720bcd5a4aed1, type: 3}
       propertyPath: m_LocalRotation.y
-      value: 0
+      value: -0
       objectReference: {fileID: 0}
     - target: {fileID: 4056836922415472167, guid: 5683d367fe60ae24586720bcd5a4aed1, type: 3}
       propertyPath: m_LocalRotation.z
-      value: 0
+      value: -0
       objectReference: {fileID: 0}
     - target: {fileID: 4056836922415472167, guid: 5683d367fe60ae24586720bcd5a4aed1, type: 3}
       propertyPath: m_LocalEulerAnglesHint.x
@@ -500,11 +495,6 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 5683d367fe60ae24586720bcd5a4aed1, type: 3}
---- !u!4 &815434203 stripped
-Transform:
-  m_CorrespondingSourceObject: {fileID: 4056836922415472167, guid: 5683d367fe60ae24586720bcd5a4aed1, type: 3}
-  m_PrefabInstance: {fileID: 815434202}
-  m_PrefabAsset: {fileID: 0}
 --- !u!1 &1006604185
 GameObject:
   m_ObjectHideFlags: 0
@@ -562,52 +552,6 @@ BoxCollider2D:
   serializedVersion: 2
   m_Size: {x: 1, y: 1}
   m_EdgeRadius: 0
---- !u!1 &1133086473
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 1133086475}
-  - component: {fileID: 1133086474}
-  m_Layer: 0
-  m_Name: FoodManager
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!114 &1133086474
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1133086473}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 93cb2ed1936920040b0d4dd5b3cbb338, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  SnakeHead: {fileID: 406708139}
-  FoodToSpawn: {fileID: 4056836922415472164, guid: 5683d367fe60ae24586720bcd5a4aed1, type: 3}
---- !u!4 &1133086475
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1133086473}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: -0.1772651}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_Children:
-  - {fileID: 815434203}
-  m_Father: {fileID: 0}
-  m_RootOrder: 7
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1157791093
 GameObject:
   m_ObjectHideFlags: 0
@@ -1007,5 +951,9 @@ PrefabInstance:
       propertyPath: m_Name
       value: Head
       objectReference: {fileID: 0}
+    - target: {fileID: 7543627872703512413, guid: 160d10cd78b53cb4081a88afe75f72f9, type: 3}
+      propertyPath: FoodToSpawn
+      value: 
+      objectReference: {fileID: 4056836922415472164, guid: 5683d367fe60ae24586720bcd5a4aed1, type: 3}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 160d10cd78b53cb4081a88afe75f72f9, type: 3}
