--- Assets/Scripts/Food/FoodSpawner.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 93cb2ed1936920040b0d4dd5b3cbb338
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
