+++ Assets/Scripts/Snake/SnakeEats.cs
@@ -7,11 +7,20 @@ public class SnakeEats : MonoBehaviour
     private GameObject FoodParticle = null;
     private GameObject[] SnakeSegments = new GameObject[] {};
 
+    [SerializeField]
+    private GameObject FoodToSpawn = null;
+
+    private List <int> AvailableSpotsX = new List<int>();
+    private List <int> AvailableSpotsY = new List<int>();
+    private Vector3 RandomPos;
+    private int RandomSelection=0;
+    private float CoordinatesPosX=0;
+    private float CoordinatesPosY=0;
+
     private float FoodPosX;
     private float FoodPosY;
     private float HeadPosX;
     private float HeadPosY;
-    bool cat = true;
 
     // Start is called before the first frame update
     void Start()
@@ -37,6 +46,24 @@ public class SnakeEats : MonoBehaviour
         if ((HeadPosX>FoodPosX-0.1f && HeadPosX<FoodPosX+0.1f ) && (HeadPosY>FoodPosY-0.1f && HeadPosY<FoodPosY+0.1f)){
             Destroy(FoodParticle);
             GridUpdate();
+
+            AvailableSpotsX.Clear();
+            AvailableSpotsY.Clear();
+            for (int j=0; j<20; j++){
+                for (int i=0; i<20; i++){
+                    if (GridStatus.GridArray[j,i]=='e'){
+                        AvailableSpotsY.Add(j);
+                        AvailableSpotsX.Add(i);
+                    }
+                }
+            }
+            RandomSelection = Random.Range(0, AvailableSpotsX.Count);
+            CoordinatesPosX = (0.4f*AvailableSpotsX[RandomSelection])-3.8f;
+            CoordinatesPosY = (-0.4f*AvailableSpotsY[RandomSelection])+2.6f;
+            RandomPos = new Vector3(CoordinatesPosX, CoordinatesPosY, 0f);
+            GameObject NewFood = Instantiate(FoodToSpawn,RandomPos,transform.rotation);
+
+
         }
     }
